generalize repl implementation
